+++ src/TelegramApi/Calendar/StringHtmlExtensions.cs
